+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Increased the RemNote bridge request timeout from 5 seconds to 15 seconds.
 - Reduced Dependabot version-update checks to monthly runs on the 15th.
 
 ## [0.13.0] - 2026-04-24